--- a/ember-cli-build.js
+++ b/ember-cli-build.js
@@ -324,6 +324,7 @@ function rollupPackage(packagesES, name) {
     'ember-debug',
     'ember-env-flags',
     'ember/features',
+    'ember/version',
     'ember-environment',
     'ember-utils',
     'ember-metal',