* Enable easy dynamic package creation (no special sauce to add a new package, just do it...)
* Enable seamless `.ts` usage throughout all of `packages/**` (just rewrite a file from `*.js` to `*.ts` and make linting happy 😝 )
* Remove arbitrary filtering/building of packages and simplify what is included and excluded from various builds.

TODO:

- [x] Remove `lib/packages.js`
- [x] Make adding new `packages/*/tests` "just work" with `each-package-tests`
- [x] Fix `node` test runs
- [x] Add `ember-utils` package to `ember-template-compiler.js` asset