@stefanpenner 2018-04-11
Can you share the before/after slow-tress output for `cold build` and `js rebuild` ?