@rwjblue 2018-04-11
<details>
<summary>The timings seem roughly unchanged (expand for specifics...)</summary>

### Before

#### Cold Initial Build

```
Build successful (77367ms) – Serving on http://localhost:7020/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
Babel (57)                                    | 67061ms (1176 ms)
TypeScriptPlugin (1)                          | 4371ms
```

#### Warm Initial Build

#### JS Rebuild

```
Build successful (2708ms) – Serving on http://localhost:7020/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
broccoli-typescript-compiler (1)              | 2376ms
```

### After

#### Cold Initial Build
```
Build successful (79002ms) – Serving on http://localhost:7020/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
Babel (7)                                     | 66325ms (9475 ms)
Rollup (21)                                   | 5227ms (248 ms)
TypeScriptPlugin (1)                          | 4704ms
```

#### Warm Initial Build

```
Build successful (12413ms) – Serving on http://localhost:7020/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
Rollup (21)                                   | 4389ms (209 ms)
TypeScriptPlugin (1)                          | 4055ms
Babel (7)                                     | 1697ms (242 ms)
Funnel (16)                                   | 898ms (56 ms)
Concat (5)                                    | 750ms (150 ms)
```

#### JS Rebuild

```
Build successful (2674ms) – Serving on http://localhost:7020/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
broccoli-typescript-compiler (1)              | 1578ms
Rollup: rollup ember-metal (1)                | 247ms
Rollup: rollup ember-glimmer (1)              | 240ms
Funnel (16)                                   | 229ms (14 ms)
```
</details>