@stefanpenner 2018-04-11
> Warm Initial Build

before: `2708ms`
after: `12413ms`


this seems like a regression?