@rwjblue 2018-04-11
The cold initial and rebuild times are roughly the same, and I didn't compare the warm initial build times after prepping the numbers. Will dig into it a bit and see what we can do to improve. 

Ultimately, IMHO the overall benefits of this PR (mentioned in the PR description) are significant and are valuable enough to justify the (hopefully temporary) warm initial build regression...