@rwjblue 2018-04-11
After starting to poke at that, I realized that the numbers you were comparing were the "Before: JS rebuild" time and "After: Warm Initial Build" time. This is because I had forgotten to paste in the "Before: Warm Initial Build" time 😭.

I've updated the comment above, but here is a screenshot of what it was before (where I accidentally trolled you 😢 ):

![original screenshot](http://assets.rwjblue.com/snapshots/2018-04-11_14-53-06_sdts4_Refactor_the_build_by_rwjblue__Pull_Request_16482__emberjsember.js.image.png)