@stefanpenner 2018-04-11
> Ultimately, I still think there is a small regression for initial builds (~ 3% on initial cold, ~ 9% on initial warm) but a small improvement for rebuilds. Definitely still room for improvement!
> 
> 😄 1

ya that type of regression isn't fatal, carry on. Thanks for digging in.